pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Kevin Jiang <git@kevinjiang.ca>"]
readme = "README.md"
packages = [{include = "api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
quart = {extras = ["dotenv"], version = "^0.19.6"}
python-dotenv = "^1.0.1"
quart-schema = "^0.20.0"
quart-cors = "^0.7.0"
openai = "^1.43.0"
uuid7 = "^0.1.0"
deepgram-sdk = "^3.5.1"
hypercorn = "^0.17.3"

[tool.poetry.scripts]
start = "api:run"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"